WIP - lots of updates for data model changes
