+++ common/models/channel.json
@@ -3,6 +3,10 @@
   "plural": "channels",
   "base": "PersistedModel",
   "idInjection": true,
+  "mixins": {
+    "Timestamp": {},
+    "IsPublicOrOwner": {}
+  },
   "options": {
     "validateUpsert": true
   },
@@ -10,6 +14,18 @@
     "name": {
       "type": "string",
       "required": true
+    },
+    "description": {
+      "type": "string",
+      "required": false
+    },
+    "is_public": {
+      "type": "boolean",
+      "required": true,
+      "default": true
+    },
+    "thumb_url": {
+      "type": "string"
     }
   },
   "validations": [],
@@ -47,7 +63,7 @@
     {
       "accessType": "READ",
       "principalType": "ROLE",
-      "principalId": "$authenticated",
+      "principalId": "$everyone",
       "permission": "ALLOW"
     },
     {
