+++ common/models/frame.json
@@ -94,6 +94,13 @@
     {
       "accessType": "EXECUTE",
       "principalType": "ROLE",
+      "principalId": "$frameManager",
+      "permission": "ALLOW",
+      "property": "update_managers_by_username"
+    },
+    {
+      "accessType": "EXECUTE",
+      "principalType": "ROLE",
       "principalId": "$owner",
       "permission": "ALLOW",
       "property": "update_managers_by_username"
