+++ common/models/openframe-user.js
@@ -151,258 +151,6 @@ module.exports = function(OpenframeUser) {
         }
     );
 
-
-
-
-
-    // Get the first collection for this user -- this will be the 'primary' collection
-    OpenframeUser.prototype.primary_collection = function(cb) {
-        var collection;
-
-        this.collections({
-            include: {
-                relation: 'artwork',
-                scope: {
-                    order: 'created DESC'
-                }
-            }
-        }, function(err, cols) {
-            collection = cols && cols.length ? cols[0] : null;
-            cb(null, collection);
-        });
-    };
-    OpenframeUser.remoteMethod(
-        'primary_collection', {
-            description: 'Get the first collection for this user.',
-            accepts: [],
-            http: {
-                verb: 'get',
-                path: '/collections/primary'
-            },
-            isStatic: false,
-            returns: {
-                arg: 'collection',
-                type: 'Array'
-            }
-        }
-    );
-
-
-
-
-
-    // Post a new artwork to this user's primary collection
-    OpenframeUser.prototype.primary_collection_new_artwork = function(artwork, cb) {
-        artwork.ownerId = this.id;
-        this.primary_collection(function(err, collection) {
-            if (collection) {
-                collection.artwork.create(artwork, function(err, artwork) {
-                    if (err) {
-                        return cb(err);
-                    }
-                    cb(null, artwork);
-                });
-            }
-        });
-    };
-    OpenframeUser.remoteMethod(
-        'primary_collection_new_artwork', {
-            description: 'Add a new artwork to the user\'s primary collection',
-            accepts: {
-                arg: 'artwork',
-                type: 'Object',
-                http: {
-                    source: 'body'
-                }
-            },
-            http: {
-                verb: 'post',
-                path: '/collections/primary/artwork'
-            },
-            isStatic: false,
-            returns: {
-                arg: 'artwork',
-                type: 'Object'
-            }
-        }
-    );
-
-
-
-
-
-    // Add an existing artwork to this user's primary collection
-    OpenframeUser.prototype.primary_collection_add_artwork = function(artworkId, cb) {
-        var Artwork = OpenframeUser.app.models.Artwork,
-            self = this;
-        Artwork.findById(artworkId, function(err, artwork) {
-            if (err) {
-                return cb(err);
-            }
-            self.primary_collection(function(err, collection) {
-                if (collection) {
-                    collection.artwork.add(artwork, function(err, artwork) {
-                        if (err) {
-                            return cb(err);
-                        }
-                        cb(null, artwork);
-                    });
-                }
-            });
-        });
-    };
-    OpenframeUser.remoteMethod(
-        'primary_collection_add_artwork', {
-            description: 'Add an existing artwork to the user\'s primary collection',
-            accepts: {
-                arg: 'artwork',
-                type: 'Object',
-                http: {
-                    source: 'body'
-                }
-            },
-            http: {
-                verb: 'put',
-                path: '/collections/primary/artwork/:artworkId'
-            },
-            isStatic: false,
-            returns: {
-                arg: 'artwork',
-                type: 'Object'
-            }
-        }
-    );
-
-
-
-
-
-    // Add an existing artwork to this user's liked_artwork and primary collection
-    OpenframeUser.prototype.like_artwork = function(artworkId, cb) {
-        var Artwork = OpenframeUser.app.models.Artwork,
-            self = this;
-
-        Artwork.findById(artworkId, function(err, artwork) {
-            if (err) {
-                return cb(err);
-            }
-
-            artwork.likers.add(self, function(err, relation) {
-                if (err) {
-                    debug(err);
-                    cb(err);
-                } else {
-                    debug(relation);
-                    self.primary_collection(function(err, collection) {
-                        if (err || !collection) {
-                            debug(err);
-                            cb(err);
-                        }
-                        if (collection) {
-                            collection.artwork.add(artwork, function(err, artwork) {
-                                if (err) {
-                                    return cb(err);
-                                }
-                                cb(null, relation);
-                            });
-                        }
-                    });
-                }
-            });
-        });
-    };
-    OpenframeUser.remoteMethod(
-        'like_artwork', {
-            description: 'Add an existing artwork to the user\'s liked_artwork and primary collection',
-            accepts: {
-                arg: 'artworkId',
-                type: 'String',
-                required: true,
-                http: {
-                    source: 'path'
-                }
-            },
-            http: {
-                verb: 'put',
-                path: '/artwork/like/:artworkId'
-            },
-            isStatic: false,
-            returns: {
-                arg: 'relation',
-                type: 'Object'
-            }
-        }
-    );
-
-
-
-
-
-    // Remove an existing artwork from this user's liked_artwork and primary collection
-    OpenframeUser.prototype.unlike_artwork = function(artworkId, cb) {
-        var Artwork = OpenframeUser.app.models.Artwork,
-            self = this;
-
-        Artwork.findById(artworkId, function(err, artwork) {
-            if (err) {
-                return cb(err);
-            }
-
-            artwork.likers.remove(self, function(err) {
-                if (err) {
-                    debug(err);
-                    cb(err);
-                } else {
-                    // if this user is the owner, can't remove from primary collection.
-                    debug(artwork.ownerId, self.id);
-                    if (artwork.ownerId === self.id) {
-                        return cb(null);
-                    }
-                    self.primary_collection(function(err, collection) {
-                        if (err || !collection) {
-                            debug(err);
-                            cb(err);
-                        }
-                        if (collection) {
-                            collection.artwork.remove(artwork, function(err) {
-                                if (err) {
-                                    return cb(err);
-                                }
-                                cb(null);
-                            });
-                        }
-                    });
-                }
-            });
-        });
-    };
-    OpenframeUser.remoteMethod(
-        'unlike_artwork', {
-            description: 'Remove an existing artwork from this user\'s liked_artwork and primary collection',
-            accepts: {
-                arg: 'artworkId',
-                type: 'String',
-                required: true,
-                http: {
-                    source: 'path'
-                }
-            },
-            http: {
-                verb: 'put',
-                path: '/artwork/unlike/:artworkId'
-            },
-            isStatic: false,
-            returns: {
-                arg: 'relation',
-                type: 'Object'
-            }
-        }
-    );
-
-
-
-
-
     /**
      * Override toJSON in order to remove inclusion of email address for users that are
      * not the currently logged-in user.
