+++ package.json
@@ -4,7 +4,7 @@
   "description": "The Openframe API Server.",
   "main": "server/server.js",
   "scripts": {
-    "start": "./bin/cli.js -p",
+    "start": "nodemon ./bin/cli.js -p",
     "start-api": "./bin/cli.js",
     "start-pubsub": "node ./node_modules/openframe-pubsubserver/start",
     "pretest": "eslint .",
@@ -61,7 +61,8 @@
     "underscore": "^1.8.3"
   },
   "devDependencies": {
-    "eslint": "^2.0.0"
+    "eslint": "^2.0.0",
+    "nodemon": "^1.9.2"
   },
   "keywords": [
     "openframe"
